api/requirements.txt
psycopg2
flask
flask_cors
sqlalchemy
python-dotenv
pandas